Fix Simspon errors.
Fix Simpson for errors involving only 2 points or 3 points given. Before, for 2 points, the 'values1' would never be created because the n=1 and the for loop would never run. Before, for 3 points, n=2 which would cause the for loop of j=2:2:n-2 to never run and so no 'values2' was created. For the 3 points, a 'values2' does not need to be calculated.